requirements.txt
torch==2.5.1
torchaudio==2.5.1
numpy>=1.24,<2.0
scipy>=1.10
pandas>=1.5,<2.0
tqdm>=4.65
soundfile>=0.12
audiomentations>=0.36
pydantic>=2.0
elevenlabs
dotenv
torchcodec
phonemizer
wordfreq
python-Levenshtein
onnxruntime==1.17.3
piper-tts==1.3.0
huggingface_hub==0.35.3

# Pin Coqui TTS and a few of its stricter deps to avoid resolver backtracking on Py 3.11
TTS==0.22.0; python_version < "3.12"
Cython==0.29.30; python_version < "3.12"
umap-learn==0.5.1; python_version < "3.12"
inflect==5.6.2; python_version < "3.12"